Backstab-Django/app/requirements.txt
requests==2.31.0
Django==4.2.3
psycopg2-binary==2.9.6
django-crontab==0.7.1
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0